+++ src/app/api/[[...route]]/images.ts
@@ -1,11 +1,12 @@
 import { Hono } from "hono";
 
 import { unsplash } from "@/lib/unsplash";
+import { verifyAuth } from "@hono/auth-js";
 
 const DEFAULT_COUNT = 50;
 const DEFAULT_COLLECTION_IDS = ["317099"];
 
-const app = new Hono().get("/", async (context) => {
+const app = new Hono().get("/", verifyAuth(), async (context) => {
   const images = await unsplash.photos.getRandom({
     collectionIds: DEFAULT_COLLECTION_IDS,
     count: DEFAULT_COUNT,
