+++ src/app/api/[[...route]]/route.ts
@@ -1,5 +1,8 @@
-import { Hono } from "hono";
+import { Context, Hono } from "hono";
 import { handle } from "hono/vercel";
+import { AuthConfig, initAuthConfig } from "@hono/auth-js";
+
+import authConfig from "@/auth.config";
 
 import ai from "./ai";
 import images from "./images";
@@ -8,7 +11,16 @@ import users from "./users";
 // Revert to "edge" if planning on running on the edge
 export const runtime = "nodejs";
 
+function getAuthConfig(context: Context): AuthConfig {
+  return {
+    secret: process.env.AUTH_SECRET,
+    ...authConfig,
+  };
+}
+
 const app = new Hono().basePath("/api");
+
+app.use("*", initAuthConfig(getAuthConfig));
 
 const routes = app
   .route("/ai", ai)
