+++ src/auth.config.ts
@@ -1,0 +1,93 @@
+import { z } from "zod";
+import bcrypt from "bcryptjs";
+import { eq } from "drizzle-orm";
+import type { NextAuthConfig } from "next-auth";
+
+import { JWT } from "next-auth/jwt";
+import Github from "next-auth/providers/github";
+import Google from "next-auth/providers/google";
+import Credentials from "next-auth/providers/credentials";
+
+import { DrizzleAdapter } from "@auth/drizzle-adapter";
+
+import { db } from "@/db/drizzle";
+import { users } from "@/db/schema";
+import { AuthConfig } from "@hono/auth-js";
+
+const CredentialsSchema = z.object({
+  email: z.string().email(),
+  password: z.string(),
+});
+
+declare module "next-auth/jwt" {
+  interface JWT {
+    id: string | undefined;
+  }
+}
+
+declare module "@auth/core/jwt" {
+  interface JWT {
+    id: string | undefined;
+  }
+}
+
+// usersTable, accountsTable, sessionsTable, verificationTokenTable 따로 지정하지 않고 기본값으로 적용
+export default {
+  adapter: DrizzleAdapter(db),
+  providers: [
+    Credentials({
+      credentials: {
+        email: { label: "Email", type: "email" },
+        password: { label: "Password", type: "password" },
+      },
+      async authorize(credentials) {
+        const validatedFields = CredentialsSchema.safeParse(credentials);
+
+        if (!validatedFields.success) return null;
+
+        const { email, password } = validatedFields.data;
+
+        // 입력된 이메일에 해당하는 유저 찾기
+        const query = await db
+          .select()
+          .from(users)
+          .where(eq(users.email, email));
+
+        const user = query[0];
+
+        if (!user || !user.password) return null;
+
+        // 입력된 비밀번호 검증
+        const passwordMatch = await bcrypt.compare(password, user.password);
+
+        if (!passwordMatch) return null;
+
+        return user;
+      },
+    }) as any,
+    Github,
+    Google,
+  ],
+  pages: {
+    signIn: "/sign-in",
+    error: "/sign-in",
+  },
+  session: {
+    strategy: "jwt",
+  },
+  callbacks: {
+    session({ session, token }) {
+      if (token.id) {
+        session.user.id = token.id;
+      }
+      return session;
+    },
+    jwt({ token, user }) {
+      if (user) {
+        token.id = user.id;
+      }
+
+      return token;
+    },
+  },
+} satisfies NextAuthConfig & AuthConfig;
